requirements.txt
torch==2.0.1
torchvision==0.15.2
diffusers==0.23.0
transformers==4.33.2
xformers==0.0.20
accelerate==0.23.0
omegaconf
clip==0.2.0
einops
kornia==0.6.12
opencv-python
opencv-contrib-python
python-gemini-api
tqdm
gradio